requirements.txt
# Core dependencies
requests>=2.28.1
beautifulsoup4>=4.11.1
feedparser>=6.0.10
pyyaml>=6.0
anthropic>=0.5.0

# Database

# Reporting
markdown>=3.4.1
jinja2>=3.1.2

# Utils
ipaddress>=1.0.23
